Basic infrastructure for face down cards in hands
Added state management functions to the card manager.
Added new deck settings and options to the deck manager
Updated the messaging system's 'is card visible' function so that GMs don't see facedown cards if they're not supposed to
Updated cards in hands to display the deck's card back if the card is face down. Currently this card back has no drag/drop functionality, it's just a token.