Added events and supporting function
Re-added the card table dictionary to save where cards are. This is to avoid having to iterate the DB list every time a delete happens; the table maps image node path and token id to the card node.

Added a token.onDelete function to pick up when card tokens are deleted and discard the card.

Updated the facedown checks in the CardManager to account for the Card Table